+++ src/Export/Data/Entity/ProductEntity.php
@@ -6,7 +6,9 @@ namespace Omikron\FactFinder\Shopware6\Export\Data\Entity;
 
 use Omikron\FactFinder\Shopware6\Export\Data\ExportEntityInterface;
 use Omikron\FactFinder\Shopware6\Export\Field\FieldInterface;
+use Shopware\Core\Content\Product\ProductCollection;
 use Shopware\Core\Content\Product\SalesChannel\SalesChannelProductEntity as Product;
+use Shopware\Core\Content\Property\Aggregate\PropertyGroupOption\PropertyGroupOptionCollection;
 
 class ProductEntity implements ExportEntityInterface
 {
@@ -24,6 +26,16 @@ class ProductEntity implements ExportEntityInterface
     public function getId(): string
     {
         return $this->product->getId();
+    }
+
+    public function getChildren(): ?ProductCollection
+    {
+        return $this->product->getChildren();
+    }
+
+    public function getProperties(): ?PropertyGroupOptionCollection
+    {
+        return $this->product->getProperties();
     }
 
     public function toArray(): array