+++ src/Export/ExportInterface.php
@@ -13,7 +13,5 @@ interface ExportInterface
 {
     public function getByContext(SalesChannelContext $context, int $batchSize = 100): iterable;
 
-    public function getCoveredEntityType(): string;
-
     public function getProducedExportEntityType(): string;
 }